+++ src/Services/Stock/Stock.Infrastructure/Persistence/Migrations/StockDbContextModelSnapshot.cs
@@ -77,12 +77,12 @@ namespace Stock.Infrastructure.Persistence.Migrations
 
             modelBuilder.Entity("Stock.Core.Models.Stock.StockPriceEntity", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint")
+                        .HasColumnType("integer")
                         .HasColumnName("id");
 
-                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<long>("Id"));
+                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<int>("Id"));
 
                     b.Property<DateTime>("CreatedAt")
                         .HasColumnType("timestamp with time zone")
